commit new update on the project structure
