+++ app/schools/[schoolCode]/parent/dashboard/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { useParams } from "next/navigation"
-import { ParentDashboard } from "@/components/parent-dashboard/parent-dashboard"
+import { ParentDashboard } from "../components/parent-dashboard"
 
 export default function ParentDashboardPage() {
   const params = useParams()
